fix: casting to long for CrossEntropyLoss in numpy DS

+++ train_colab_152m_resume.py
@@ -156,7 +156,7 @@ class DS(torch.utils.data.Dataset):
         self.y = y
     def __len__(self): return len(self.x)
     def __getitem__(self, i):
-        return torch.from_numpy(self.x[i]), torch.from_numpy(self.y[i])
+        return torch.from_numpy(self.x[i]).long(), torch.from_numpy(self.y[i]).long()
 
 log("Building samples (pre-tokenization)...")
 if os.path.exists(NPZ):
